feature: All breakpoints type creation UI

+++ src/Spice86/ViewModels/DisassemblyViewModel.cs
@@ -63,7 +63,7 @@ public partial class DisassemblyViewModel : ViewModelWithErrorDialog {
     public State State => _state;
 
     private void OnBreakPointUpdateFromBreakpointsViewModel(BreakpointViewModel breakpointViewModel) {
-        UpdateAssemblyLineIfShown(breakpointViewModel.Address, breakpointViewModel);
+        UpdateAssemblyLineIfShown((uint)breakpointViewModel.Address, breakpointViewModel);
     }
 
     private void OnResumed() {
